requirements.txt
pandas
sqlalchemy
pymysql
python-dotenv
matplotlib
seaborn